+++ Cpp/Executables/Generate/SIR_Sine_Discrete.cpp
@@ -25,21 +25,21 @@ sineparam_gen(uint16_t
               seed) {
     std::mt19937 rng(seed);
     FROLS::Integrators::SIR_Sine_Param p;
-    double R0 = 2;
+    float R0 = 2;
     p.
             alpha = .11;
     p.N_pop = 1000;
 
     p.
             beta = R0*p.alpha/p.N_pop;
-    std::uniform_real_distribution<double> d_u((2 * M_PI) / 100, (2 * M_PI) / 1000);
+    std::uniform_real_distribution<float> d_u((2 * M_PI) / 100, (2 * M_PI) / 1000);
     p.
             omega = d_u(rng);
-    std::uniform_real_distribution<double> d_pi(0, 2 * M_PI);
+    std::uniform_real_distribution<float> d_pi(0, 2 * M_PI);
     p.
             offset = d_pi(rng);
 
-    std::uniform_real_distribution<double> d_amp(p.beta / 5, p.beta / 1);
+    std::uniform_real_distribution<float> d_amp(p.beta / 5, p.beta / 1);
     p.
             amplitude = d_amp(rng);
     return
@@ -51,27 +51,27 @@ int main() {
     using namespace FROLS;
     using namespace FROLS::Integrators;
     using Trajectory = SIR_Sine<Nt>::Trajectory;
-    double dt = 2.0;
+    float dt = 2.0;
 
     uint16_t N_sim = 10000;
     std::vector<uint16_t> seeds(N_sim);
     std::random_device rd;
     std::generate(seeds.begin(), seeds.end(), [&rd]() { return rd(); });
     std::vector<std::string> colnames = {"S", "I", "R"};
-    typedef std::vector<std::vector<double>> vec_Trajectory;
+    typedef std::vector<std::vector<float>> vec_Trajectory;
     std::for_each(seeds.begin(), seeds.end(), [&](const uint16_t seed) {
         static int iter = 0;
         SIR_Sine_Param p = sineparam_gen(seed);
-        const std::array<double, 3> x0 = {p.N_pop*9./10, p.N_pop/10, 0};
+        const std::array<float, 3> x0 = {p.N_pop*9.f/10.f, p.N_pop/10.f, 0.f};
         Integrators::SIR_Discrete<Nt> model(p);
         auto [traj, t] = model.simulate(x0);
-        std::vector<std::vector<double>> result(Nt + 1);
+        std::vector<std::vector<float>> result(Nt + 1);
         DataFrame df;
         auto p_I = t;
         std::transform(t.begin(), t.end(), p_I.begin(),
                        [&](auto ti) { return FROLS::Integrators::SIR_sine_param_gen((void *) &p, ti).beta; });
         df.assign("p_I", p_I);
-        df.assign("t", std::vector<double>(t.data(), t.end()));
+        df.assign("t", std::vector<float>(t.data(), t.end()));
         df.assign(colnames, traj);
         df.write_csv(SIR_Sine_Discrete_filename(iter), ",");
         iter++;
