+++ Cpp/Executables/Regression/truncate.cpp
@@ -7,7 +7,7 @@ int main()
 {
     uint16_t N_sims = 10000;
     uint16_t N_pop = 60;
-    double p_ER = 1.0;
+    float p_ER = 1.0;
     using namespace FROLS;
     std::vector<std::string> df_names(N_sims);
 
@@ -18,7 +18,7 @@ int main()
 
     DataFrameStack dfs(df_names);
     // Mat X_raw = dataframe_to_matrix(dfs[0], {"S", "I", "R", "p_I"})(Eigen::seq(0, Eigen::last - 1), Eigen::all);
-    double threshold = 1e-1;
+    float threshold = 1e-1;
 
     Mat X_trunc = dmd_truncate(dfs, {"S", "I", "R"}, threshold); 
     std::cout << X_trunc.topRows(3) << std::endl;
